checkout: include the list refresh into the checkout/extend session actions
If a user clicks on one of the two actions then we execute that action and
refresh the list right away. This way there is only one visible change
and no extra loading indicator required.

The only time we display an extra loading spinner now is on the very first
list fetch.